feat(like): 좋아요 상품 페이징 목록 진행중

+++ src/main/java/com/nhnacademy/front/product/like/controller/LikeController.java
@@ -2,6 +2,7 @@ package com.nhnacademy.front.product.like.controller;
 
 import java.util.Objects;
 
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
@@ -15,10 +16,13 @@ import org.springframework.web.bind.annotation.PostMapping;
 
 import com.nhnacademy.front.common.annotation.JwtTokenCheck;
 import com.nhnacademy.front.common.page.PageResponse;
+import com.nhnacademy.front.common.page.PageResponseConverter;
+import com.nhnacademy.front.jwt.parser.JwtGetMemberId;
 import com.nhnacademy.front.product.like.model.dto.request.RequestCreateLikeDTO;
 import com.nhnacademy.front.product.like.model.dto.response.ResponseLikedProductDTO;
 import com.nhnacademy.front.product.like.service.LikeService;
 
+import io.jsonwebtoken.JwtException;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 
@@ -37,11 +41,10 @@ public class LikeController {
 			throw new IllegalArgumentException();
 		}
 
-		// String memberId = JwtGetMemberId.jwtGetMemberId(request);
-		// if (Objects.isNull(memberId)) {
-		// 	throw new JwtException(JWT_TOKEN_NULL_MESSAGE);
-		// }
-		String memberId = "user";
+		String memberId = JwtGetMemberId.jwtGetMemberId(request);
+		if (Objects.isNull(memberId)) {
+			throw new JwtException(JWT_TOKEN_NULL_MESSAGE);
+		}
 
 		RequestCreateLikeDTO requestDto = new RequestCreateLikeDTO();
 		requestDto.setMemberId(memberId);
@@ -57,11 +60,10 @@ public class LikeController {
 			throw new IllegalArgumentException();
 		}
 
-		// String memberId = JwtGetMemberId.jwtGetMemberId(request);
-		// if (Objects.isNull(memberId)) {
-		// 	throw new JwtException(JWT_TOKEN_NULL_MESSAGE);
-		// }
-		String memberId = "user";
+		String memberId = JwtGetMemberId.jwtGetMemberId(request);
+		if (Objects.isNull(memberId)) {
+			throw new JwtException(JWT_TOKEN_NULL_MESSAGE);
+		}
 
 		likeService.deleteLike(productId, memberId);
 		return ResponseEntity.ok().build();
@@ -69,14 +71,15 @@ public class LikeController {
 
 	@JwtTokenCheck
 	@GetMapping("/products/likes")
-	public String getLikedProductsByCustomer(@PageableDefault(size = 8, sort = "likeCreatedAt", direction = Sort.Direction.DESC) Pageable pageable, HttpServletRequest request, Model model) {
-		// String memberId = JwtGetMemberId.jwtGetMemberId(request);
-		// if (Objects.isNull(memberId)) {
-		// 	throw new JwtException(JWT_TOKEN_NULL_MESSAGE);
-		// }
-		String memberId = "user";
+	public String getLikedProductsByCustomer(@PageableDefault(size = 6, sort = "likeCreatedAt", direction = Sort.Direction.DESC) Pageable pageable, HttpServletRequest request, Model model) {
+		String memberId = JwtGetMemberId.jwtGetMemberId(request);
+		if (Objects.isNull(memberId)) {
+			throw new JwtException(JWT_TOKEN_NULL_MESSAGE);
+		}
 
-		PageResponse<ResponseLikedProductDTO> likeProductsByCustomer = likeService.getLikeProductsByCustomer(memberId, pageable);
+		PageResponse<ResponseLikedProductDTO> response = likeService.getLikeProductsByCustomer(memberId, pageable);
+		Page<ResponseLikedProductDTO> likeProductsByCustomer = PageResponseConverter.toPage(response);
+
 		model.addAttribute("likeProducts", likeProductsByCustomer);
 
 		return "product/products-like";
